backend/TripMateFunctions/serializers/f7_2_serializers.py
from rest_framework import serializers


class F72LandingContentSerializer(serializers.Serializer):
    features = serializers.ListField(child=serializers.CharField())
    faqs = serializers.ListField(child=serializers.CharField())
